requirements.txt
psycopg2-binary
pydantic
uvicorn~=0.30.5
fastapi~=0.112.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.32
pytest~=8.3.2